Fixes for performance
